refactor: Remove commented out code in index.js and Blogs-three-column2 component
This commit removes commented out code in the index.js file and the Blogs-three-column2 component. The commented out code for the PortfolioCustomColumn and Testimonials1 components in the index.js file is no longer needed and can be safely removed. Similarly, the commented out code for the second and third blog items in the Blogs-three-column2 component is no longer needed and can be removed. This cleanup improves code readability and reduces clutter.